requirements.txt
asgiref==3.2.10
certifi==2020.6.20
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1601563330930/work
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1594305108220/work
pytz==2020.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
whitenoise==5.2.0
